# specs/openapi/task-service.yaml
openapi: "3.1.0"

info:
  title: Task Service API
  description: タスク管理サービスの REST API
  version: "0.1.0"

servers:
  - url: http://localhost:3000
    description: ローカル開発環境

tags:
  - name: tasks
    description: タスク操作

paths:
  /api/tasks:
    get:
      operationId: listTasks
      summary: タスク一覧を取得する
      tags: [tasks]
      parameters:
        - name: status
          in: query
          required: false
          description: ステータスで絞り込む
          schema:
            $ref: "#/components/schemas/TaskStatus"
      responses:
        "200":
          description: タスク一覧
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"

    post:
      operationId: createTask
      summary: タスクを作成する
      tags: [tasks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskInput"
      responses:
        "201":
          description: 作成されたタスク
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "422":
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"

  /api/tasks/{taskId}:
    parameters:
      - name: taskId
        in: path
        required: true
        description: タスク ID (UUID)
        schema:
          type: string
          format: uuid

    get:
      operationId: getTask
      summary: タスクを1件取得する
      tags: [tasks]
      responses:
        "200":
          description: タスク
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: タスクが存在しない
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"

    patch:
      operationId: updateTask
      summary: タスクを更新する
      tags: [tasks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskInput"
      responses:
        "200":
          description: 更新後のタスク
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: タスクが存在しない
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
        "422":
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"

    delete:
      operationId: deleteTask
      summary: タスクを削除する
      tags: [tasks]
      responses:
        "200":
          description: 削除成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteTaskResult"
        "404":
          description: タスクが存在しない
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"

components:
  schemas:
    TaskStatus:
      type: string
      enum: [todo, in_progress, done]
      description: |
        - `todo`: 未着手
        - `in_progress`: 進行中
        - `done`: 完了

    Task:
      type: object
      required:
        - id
        - title
        - description
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          description: タスク ID
          examples: ["550e8400-e29b-41d4-a716-446655440000"]
        title:
          type: string
          minLength: 1
          maxLength: 255
          description: タスクのタイトル
          examples: ["仕様書を書く"]
        description:
          type: string
          nullable: true
          description: タスクの詳細説明（任意）
          examples: ["OpenAPI スキーマを作成する"]
        status:
          $ref: "#/components/schemas/TaskStatus"
        createdAt:
          type: string
          format: date-time
          description: 作成日時 (ISO 8601)
          examples: ["2026-02-19T00:00:00.000Z"]
        updatedAt:
          type: string
          format: date-time
          description: 最終更新日時 (ISO 8601)
          examples: ["2026-02-19T00:00:00.000Z"]

    CreateTaskInput:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
          description: タスクのタイトル
          examples: ["仕様書を書く"]
        description:
          type: string
          description: タスクの詳細説明（任意）
          examples: ["OpenAPI スキーマを作成する"]

    UpdateTaskInput:
      type: object
      minProperties: 1
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
          description: タスクのタイトル
        description:
          type: string
          nullable: true
          description: タスクの詳細説明（null で削除）
        status:
          $ref: "#/components/schemas/TaskStatus"

    DeleteTaskResult:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          enum: [true]

    NotFoundError:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          examples: ["Task not found"]

    ValidationError:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          examples: ["Validation failed"]
        details:
          type: array
          items:
            type: object
            required:
              - field
              - message
            properties:
              field:
                type: string
              message:
                type: string